+++ x/ibc/core/keeper/msg_server.go
@@ -3,10 +3,15 @@ package keeper
 import (
 	"context"
 
+	"github.com/armon/go-metrics"
+	"github.com/cosmos/cosmos-sdk/telemetry"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
 	clienttypes "github.com/cosmos/cosmos-sdk/x/ibc/core/02-client/types"
 	connectiontypes "github.com/cosmos/cosmos-sdk/x/ibc/core/03-connection/types"
+	channel "github.com/cosmos/cosmos-sdk/x/ibc/core/04-channel"
+	channeltypes "github.com/cosmos/cosmos-sdk/x/ibc/core/04-channel/types"
+	porttypes "github.com/cosmos/cosmos-sdk/x/ibc/core/05-port/types"
 )
 
 type MsgServer struct {
@@ -240,4 +245,363 @@ func (k MsgServer) ConnectionOpenConfirm(goCtx context.Context, msg *connectiont
 	})
 
 	return &connectiontypes.MsgConnectionOpenConfirmResponse{}, nil
+}
+
+func (k MsgServer) ChannelOpenInit(goCtx context.Context, msg *channeltypes.MsgChannelOpenInit) (*channeltypes.MsgChannelOpenInitResponse, error) {
+	ctx := sdk.UnwrapSDKContext(goCtx)
+
+	// Lookup module by port capability
+	module, portCap, err := k.PortKeeper.LookupModuleByPort(ctx, msg.PortId)
+	if err != nil {
+		return nil, sdkerrors.Wrap(err, "could not retrieve module from port-id")
+	}
+
+	_, cap, err := channel.HandleMsgChannelOpenInit(ctx, k.ChannelKeeper, portCap, msg)
+	if err != nil {
+		return nil, err
+	}
+
+	// Retrieve callbacks from router
+	cbs, ok := k.Router.GetRoute(module)
+	if !ok {
+		return nil, sdkerrors.Wrapf(porttypes.ErrInvalidRoute, "route not found to module: %s", module)
+	}
+
+	if err = cbs.OnChanOpenInit(ctx, msg.Channel.Ordering, msg.Channel.ConnectionHops, msg.PortId, msg.ChannelId, cap, msg.Channel.Counterparty, msg.Channel.Version); err != nil {
+		return nil, sdkerrors.Wrap(err, "channel open init callback failed")
+	}
+
+	return &channeltypes.MsgChannelOpenInitResponse{}, nil
+}
+
+func (k MsgServer) ChannelOpenTry(goCtx context.Context, msg *channeltypes.MsgChannelOpenTry) (*channeltypes.MsgChannelOpenTryResponse, error) {
+	ctx := sdk.UnwrapSDKContext(goCtx)
+	// Lookup module by port capability
+	module, portCap, err := k.PortKeeper.LookupModuleByPort(ctx, msg.PortId)
+	if err != nil {
+		return nil, sdkerrors.Wrap(err, "could not retrieve module from port-id")
+	}
+
+	_, cap, err := channel.HandleMsgChannelOpenTry(ctx, k.ChannelKeeper, portCap, msg)
+	if err != nil {
+		return nil, err
+	}
+
+	// Retrieve callbacks from router
+	cbs, ok := k.Router.GetRoute(module)
+	if !ok {
+		return nil, sdkerrors.Wrapf(porttypes.ErrInvalidRoute, "route not found to module: %s", module)
+	}
+
+	if err = cbs.OnChanOpenTry(ctx, msg.Channel.Ordering, msg.Channel.ConnectionHops, msg.PortId, msg.DesiredChannelId, cap, msg.Channel.Counterparty, msg.Channel.Version, msg.CounterpartyVersion); err != nil {
+		return nil, sdkerrors.Wrap(err, "channel open try callback failed")
+	}
+
+	return &channeltypes.MsgChannelOpenTryResponse{}, nil
+}
+
+func (k MsgServer) ChannelOpenAck(goCtx context.Context, msg *channeltypes.MsgChannelOpenAck) (*channeltypes.MsgChannelOpenAckResponse, error) {
+	ctx := sdk.UnwrapSDKContext(goCtx)
+
+	// Lookup module by channel capability
+	module, cap, err := k.ChannelKeeper.LookupModuleByChannel(ctx, msg.PortId, msg.ChannelId)
+	if err != nil {
+		return nil, sdkerrors.Wrap(err, "could not retrieve module from port-id")
+	}
+
+	// Retrieve callbacks from router
+	cbs, ok := k.Router.GetRoute(module)
+	if !ok {
+		return nil, sdkerrors.Wrapf(porttypes.ErrInvalidRoute, "route not found to module: %s", module)
+	}
+
+	if err = cbs.OnChanOpenAck(ctx, msg.PortId, msg.ChannelId, msg.CounterpartyVersion); err != nil {
+		return nil, sdkerrors.Wrap(err, "channel open ack callback failed")
+	}
+
+	_, err = channel.HandleMsgChannelOpenAck(ctx, k.ChannelKeeper, cap, msg)
+	if err != nil {
+		return nil, err
+	}
+
+	return &channeltypes.MsgChannelOpenAckResponse{}, nil
+}
+
+func (k MsgServer) ChannelOpenConfirm(goCtx context.Context, msg *channeltypes.MsgChannelOpenConfirm) (*channeltypes.MsgChannelOpenConfirmResponse, error) {
+	ctx := sdk.UnwrapSDKContext(goCtx)
+
+	// Lookup module by channel capability
+	module, cap, err := k.ChannelKeeper.LookupModuleByChannel(ctx, msg.PortId, msg.ChannelId)
+	if err != nil {
+		return nil, sdkerrors.Wrap(err, "could not retrieve module from port-id")
+	}
+
+	// Retrieve callbacks from router
+	cbs, ok := k.Router.GetRoute(module)
+	if !ok {
+		return nil, sdkerrors.Wrapf(porttypes.ErrInvalidRoute, "route not found to module: %s", module)
+	}
+
+	if err = cbs.OnChanOpenConfirm(ctx, msg.PortId, msg.ChannelId); err != nil {
+		return nil, sdkerrors.Wrap(err, "channel open confirm callback failed")
+	}
+
+	_, err = channel.HandleMsgChannelOpenConfirm(ctx, k.ChannelKeeper, cap, msg)
+	if err != nil {
+		return nil, err
+	}
+
+	return &channeltypes.MsgChannelOpenConfirmResponse{}, nil
+}
+
+func (k MsgServer) ChannelCloseInit(goCtx context.Context, msg *channeltypes.MsgChannelCloseInit) (*channeltypes.MsgChannelCloseInitResponse, error) {
+	ctx := sdk.UnwrapSDKContext(goCtx)
+	// Lookup module by channel capability
+	module, cap, err := k.ChannelKeeper.LookupModuleByChannel(ctx, msg.PortId, msg.ChannelId)
+	if err != nil {
+		return nil, sdkerrors.Wrap(err, "could not retrieve module from port-id")
+	}
+
+	// Retrieve callbacks from router
+	cbs, ok := k.Router.GetRoute(module)
+	if !ok {
+		return nil, sdkerrors.Wrapf(porttypes.ErrInvalidRoute, "route not found to module: %s", module)
+	}
+
+	if err = cbs.OnChanCloseInit(ctx, msg.PortId, msg.ChannelId); err != nil {
+		return nil, sdkerrors.Wrap(err, "channel close init callback failed")
+	}
+
+	_, err = channel.HandleMsgChannelCloseInit(ctx, k.ChannelKeeper, cap, msg)
+	if err != nil {
+		return nil, err
+	}
+
+	return &channeltypes.MsgChannelCloseInitResponse{}, nil
+}
+
+func (k MsgServer) ChannelCloseConfirm(goCtx context.Context, msg *channeltypes.MsgChannelCloseConfirm) (*channeltypes.MsgChannelCloseConfirmResponse, error) {
+	ctx := sdk.UnwrapSDKContext(goCtx)
+
+	// Lookup module by channel capability
+	module, cap, err := k.ChannelKeeper.LookupModuleByChannel(ctx, msg.PortId, msg.ChannelId)
+	if err != nil {
+		return nil, sdkerrors.Wrap(err, "could not retrieve module from port-id")
+	}
+
+	// Retrieve callbacks from router
+	cbs, ok := k.Router.GetRoute(module)
+	if !ok {
+		return nil, sdkerrors.Wrapf(porttypes.ErrInvalidRoute, "route not found to module: %s", module)
+	}
+
+	if err = cbs.OnChanCloseConfirm(ctx, msg.PortId, msg.ChannelId); err != nil {
+		return nil, sdkerrors.Wrap(err, "channel close confirm callback failed")
+	}
+
+	_, err = channel.HandleMsgChannelCloseConfirm(ctx, k.ChannelKeeper, cap, msg)
+	if err != nil {
+		return nil, err
+	}
+
+	return &channeltypes.MsgChannelCloseConfirmResponse{}, nil
+}
+
+func (k MsgServer) RecvPacket(goCtx context.Context, msg *channeltypes.MsgRecvPacket) (*channeltypes.MsgRecvPacketResponse, error) {
+	ctx := sdk.UnwrapSDKContext(goCtx)
+
+	// Lookup module by channel capability
+	module, cap, err := k.ChannelKeeper.LookupModuleByChannel(ctx, msg.Packet.DestinationPort, msg.Packet.DestinationChannel)
+	if err != nil {
+		return nil, sdkerrors.Wrap(err, "could not retrieve module from port-id")
+	}
+
+	// Retrieve callbacks from router
+	cbs, ok := k.Router.GetRoute(module)
+	if !ok {
+		return nil, sdkerrors.Wrapf(porttypes.ErrInvalidRoute, "route not found to module: %s", module)
+	}
+
+	// Perform TAO verification
+	if err := k.ChannelKeeper.RecvPacket(ctx, msg.Packet, msg.Proof, msg.ProofHeight); err != nil {
+		return nil, sdkerrors.Wrap(err, "receive packet verification failed")
+	}
+
+	// Perform application logic callback
+	_, ack, err := cbs.OnRecvPacket(ctx, msg.Packet)
+	if err != nil {
+		return nil, sdkerrors.Wrap(err, "receive packet callback failed")
+	}
+
+	if err := k.ChannelKeeper.WriteReceipt(ctx, cap, msg.Packet); err != nil {
+		return nil, err
+	}
+
+	// Set packet acknowledgement only if the acknowledgement is not nil.
+	// NOTE: IBC applications modules may call the WriteAcknowledgement asynchronously if the
+	// acknowledgement is nil.
+	if ack != nil {
+		if err := k.ChannelKeeper.WriteAcknowledgement(ctx, msg.Packet, ack); err != nil {
+			return nil, err
+		}
+	}
+
+	defer func() {
+		telemetry.IncrCounterWithLabels(
+			[]string{"tx", "msg", "ibc", msg.Type()},
+			1,
+			[]metrics.Label{
+				telemetry.NewLabel("source-port", msg.Packet.SourcePort),
+				telemetry.NewLabel("source-channel", msg.Packet.SourceChannel),
+				telemetry.NewLabel("destination-port", msg.Packet.DestinationPort),
+				telemetry.NewLabel("destination-channel", msg.Packet.DestinationChannel),
+			},
+		)
+	}()
+
+	return &channeltypes.MsgRecvPacketResponse{}, nil
+}
+
+func (k MsgServer) Timeout(goCtx context.Context, msg *channeltypes.MsgTimeout) (*channeltypes.MsgTimeoutResponse, error) {
+	ctx := sdk.UnwrapSDKContext(goCtx)
+	// Lookup module by channel capability
+	module, cap, err := k.ChannelKeeper.LookupModuleByChannel(ctx, msg.Packet.SourcePort, msg.Packet.SourceChannel)
+	if err != nil {
+		return nil, sdkerrors.Wrap(err, "could not retrieve module from port-id")
+	}
+
+	// Retrieve callbacks from router
+	cbs, ok := k.Router.GetRoute(module)
+	if !ok {
+		return nil, sdkerrors.Wrapf(porttypes.ErrInvalidRoute, "route not found to module: %s", module)
+	}
+
+	// Perform TAO verification
+	if err := k.ChannelKeeper.TimeoutPacket(ctx, msg.Packet, msg.Proof, msg.ProofHeight, msg.NextSequenceRecv); err != nil {
+		return nil, sdkerrors.Wrap(err, "timeout packet verification failed")
+	}
+
+	// Perform application logic callback
+	_, err = cbs.OnTimeoutPacket(ctx, msg.Packet)
+	if err != nil {
+		return nil, sdkerrors.Wrap(err, "timeout packet callback failed")
+	}
+
+	// Delete packet commitment
+	if err = k.ChannelKeeper.TimeoutExecuted(ctx, cap, msg.Packet); err != nil {
+		return nil, err
+	}
+
+	defer func() {
+		telemetry.IncrCounterWithLabels(
+			[]string{"ibc", "timeout", "packet"},
+			1,
+			[]metrics.Label{
+				telemetry.NewLabel("source-port", msg.Packet.SourcePort),
+				telemetry.NewLabel("source-channel", msg.Packet.SourceChannel),
+				telemetry.NewLabel("destination-port", msg.Packet.DestinationPort),
+				telemetry.NewLabel("destination-channel", msg.Packet.DestinationChannel),
+				telemetry.NewLabel("timeout-type", "height"),
+			},
+		)
+	}()
+
+	return &channeltypes.MsgTimeoutResponse{}, nil
+}
+
+func (k MsgServer) TimeoutOnClose(goCtx context.Context, msg *channeltypes.MsgTimeoutOnClose) (*channeltypes.MsgTimeoutOnCloseResponse, error) {
+	ctx := sdk.UnwrapSDKContext(goCtx)
+
+	// Lookup module by channel capability
+	module, cap, err := k.ChannelKeeper.LookupModuleByChannel(ctx, msg.Packet.SourcePort, msg.Packet.SourceChannel)
+	if err != nil {
+		return nil, sdkerrors.Wrap(err, "could not retrieve module from port-id")
+	}
+
+	// Retrieve callbacks from router
+	cbs, ok := k.Router.GetRoute(module)
+	if !ok {
+		return nil, sdkerrors.Wrapf(porttypes.ErrInvalidRoute, "route not found to module: %s", module)
+	}
+
+	// Perform TAO verification
+	if err := k.ChannelKeeper.TimeoutOnClose(ctx, cap, msg.Packet, msg.Proof, msg.ProofClose, msg.ProofHeight, msg.NextSequenceRecv); err != nil {
+		return nil, sdkerrors.Wrap(err, "timeout on close packet verification failed")
+	}
+
+	// Perform application logic callback
+	// NOTE: MsgTimeout and MsgTimeoutOnClose use the same "OnTimeoutPacket"
+	// application logic callback.
+	_, err = cbs.OnTimeoutPacket(ctx, msg.Packet)
+	if err != nil {
+		return nil, sdkerrors.Wrap(err, "timeout packet callback failed")
+	}
+
+	// Delete packet commitment
+	if err = k.ChannelKeeper.TimeoutExecuted(ctx, cap, msg.Packet); err != nil {
+		return nil, err
+	}
+
+	defer func() {
+		telemetry.IncrCounterWithLabels(
+			[]string{"ibc", "timeout", "packet"},
+			1,
+			[]metrics.Label{
+				telemetry.NewLabel("source-port", msg.Packet.SourcePort),
+				telemetry.NewLabel("source-channel", msg.Packet.SourceChannel),
+				telemetry.NewLabel("destination-port", msg.Packet.DestinationPort),
+				telemetry.NewLabel("destination-channel", msg.Packet.DestinationChannel),
+				telemetry.NewLabel("timeout-type", "channel-closed"),
+			},
+		)
+	}()
+
+	return &channeltypes.MsgTimeoutOnCloseResponse{}, nil
+}
+
+func (k MsgServer) Acknowledgement(goCtx context.Context, msg *channeltypes.MsgAcknowledgement) (*channeltypes.MsgAcknowledgementResponse, error) {
+	ctx := sdk.UnwrapSDKContext(goCtx)
+
+	// Lookup module by channel capability
+	module, cap, err := k.ChannelKeeper.LookupModuleByChannel(ctx, msg.Packet.SourcePort, msg.Packet.SourceChannel)
+	if err != nil {
+		return nil, sdkerrors.Wrap(err, "could not retrieve module from port-id")
+	}
+
+	// Retrieve callbacks from router
+	cbs, ok := k.Router.GetRoute(module)
+	if !ok {
+		return nil, sdkerrors.Wrapf(porttypes.ErrInvalidRoute, "route not found to module: %s", module)
+	}
+
+	// Perform TAO verification
+	if err := k.ChannelKeeper.AcknowledgePacket(ctx, msg.Packet, msg.Acknowledgement, msg.Proof, msg.ProofHeight); err != nil {
+		return nil, sdkerrors.Wrap(err, "acknowledge packet verification failed")
+	}
+
+	// Perform application logic callback
+	_, err = cbs.OnAcknowledgementPacket(ctx, msg.Packet, msg.Acknowledgement)
+	if err != nil {
+		return nil, sdkerrors.Wrap(err, "acknowledge packet callback failed")
+	}
+
+	// Delete packet commitment
+	if err = k.ChannelKeeper.AcknowledgementExecuted(ctx, cap, msg.Packet); err != nil {
+		return nil, err
+	}
+
+	defer func() {
+		telemetry.IncrCounterWithLabels(
+			[]string{"tx", "msg", "ibc", msg.Type()},
+			1,
+			[]metrics.Label{
+				telemetry.NewLabel("source-port", msg.Packet.SourcePort),
+				telemetry.NewLabel("source-channel", msg.Packet.SourceChannel),
+				telemetry.NewLabel("destination-port", msg.Packet.DestinationPort),
+				telemetry.NewLabel("destination-channel", msg.Packet.DestinationChannel),
+			},
+		)
+	}()
+
+	return &channeltypes.MsgAcknowledgementResponse{}, nil
 }
